+++ IPX800cs/Parsers/IAnalogInputResponseParser.cs
@@ -2,6 +2,6 @@ namespace IPX800cs.Parsers
 {
     public interface IAnalogInputResponseParser
     {
-        double ParseResponse(string ipxResponse, int inputNumber);
+        int ParseResponse(string ipxResponse, int inputNumber);
     }
 }
