+++ IPX800cs/Parsers/v3/Legacy/M2M/IPX800v3LegacyGetAnalogInputM2MResponseParser.cs
@@ -2,10 +2,10 @@
 {
     internal class IPX800v3LegacyGetAnalogInputM2MResponseParser : IAnalogInputResponseParser
     {
-        public double ParseResponse(string ipxResponse, int inputNumber)
+        public int ParseResponse(string ipxResponse, int inputNumber)
         {
             var response = ipxResponse.Trim().Split('=');
-            return double.Parse(response[1]);
+            return int.Parse(response[1]);
         }
     }
 }
